requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.2.1
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
factory-boy==2.10.0
Faker==0.8.12
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
